Allow reusing integration test environment
This updates the run.sh script in the integration test to check for
the presence of the REUSE_TEST_ENV environment variable.

If present, the containers and volumes for MinIO and MySQL will be
kept which allows faster startup time between different test runs.

+++ tests/integration/run.sh
@@ -16,6 +16,8 @@ docker compose run --rm archivematica-storage-service
 
 status=$?
 
-docker compose down --volumes
+if [ -z "${REUSE_TEST_ENV}" ]; then
+    docker compose down --volumes
+fi
 
 exit $status
